requirements.txt
absl-py==2.1.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
Django==2.1.15
django-cors-headers==4.3.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
executing==2.0.1
filelock==3.13.4
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
mysqlclient==2.2.1
namex==0.0.8
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pywin32==306
pyzmq==26.0.3
rating==0.1.0
requests==2.31.0
rich==13.7.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.1
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
azure-core==1.30.1
azure-storage-blob==12.19.1
